# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SORRISO PERFEITO API",
    "description": "API interativa para gestão de clínica odontológica, incluindo controle de acessos (usuários), pacientes, prontuário (anamnese), odontólogos, agendamento de consultas e catalogação de procedimentos clínicos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/consultas": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Listar consultas agendadas",
        "responses": {
          "200": {
            "description": "Lista de consultas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Consulta"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Agendar uma nova consulta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consulta marcada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/consultas/{id}": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Buscar consulta por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da consulta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "$ref": "#/components/schemas/Consulta"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consulta não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Consultas"
        ],
        "summary": "Atualizar dados de uma consulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consulta atualizada"
          }
        }
      },
      "delete": {
        "tags": [
          "Consultas"
        ],
        "summary": "Excluir agendamento de consulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta excluída"
          }
        }
      }
    },
    "/consultas/{id_consulta}/procedimentos": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Listar procedimentos realizados na consulta",
        "parameters": [
          {
            "name": "id_consulta",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de procedimentos associados à consulta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer", "example": 1 },
                          "nome": { "type": "string", "example": "Limpeza Dental" },
                          "valor": { "type": "number", "format": "float", "example": 150.00 }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Vincular um procedimento realizado à consulta",
        "parameters": [
          {
            "name": "id_consulta",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id_procedimento"
                ],
                "properties": {
                  "id_procedimento": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Procedimento vinculado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Procedimento já vinculado à consulta"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponseEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Operação executada com sucesso"
              }
            }
          }
        }
      },
...
      "Consulta": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "data_hora": { "type": "string", "format": "date-time", "example": "2026-05-25T10:00:00" },
          "motivo": { "type": "string", "example": "Rotina Geral" },
          "valor": { "type": "number", "format": "float", "example": 150.00 },
          "prioridade": { "type": "string", "enum": ["BAIXA", "MEDIA", "ALTA"], "example": "MEDIA" },
          "paciente": { "type": "string", "example": "Ana Santos" },
          "odontologo": { "type": "string", "example": "Dr. Fernando Dentista" },
          "usuario_responsavel": { "type": "string", "example": "Recepcionista Maria" }
        }
      },
      "ConsultaInput": {
        "type": "object",
        "required": [
          "id_paciente",
          "id_odontologo",
          "id_usuario_responsavel",
          "data_hora",
          "valor"
        ],
        "properties": {
          "id_paciente": { "type": "integer", "example": 5 },
          "id_odontologo": { "type": "integer", "example": 2 },
          "id_usuario_responsavel": { "type": "integer", "example": 1 },
          "data_hora": { "type": "string", "format": "date-time", "example": "2026-05-25T10:00:00" },
          "motivo": { "type": "string", "example": "Rotina Geral" },
          "valor": { "type": "number", "minimum": 0, "example": 150.00 },
          "prioridade": { "type": "string", "enum": ["BAIXA", "MEDIA", "ALTA"], "example": "MEDIA" }
        }
      },
...
    }
  }
}